+++ changelog.md
@@ -21,6 +21,7 @@
 ### Enhancements
 * `SetDirection.run` now supports `ConductingEquipment`.
 * Fixed types on all overrides for `PowerSystemResource.asset_info`, removing the need to shadow them with type specific variants.
+* You can now pass a list of `TransformerEndRatedS` to the `PowerTransformerEnd` constructor via the `ratings` argument.
 
 ### Fixes
 * None.